+++ src/utils/LazyLoader.js
@@ -7,8 +7,8 @@ export default {
         import(/* webpackChunkName: "dropdown" */ '../fragments/Dropdown.react'),
     graph: () =>
         import(/* webpackChunkName: "graph" */ '../fragments/Graph.react'),
-    // markdown: () =>
-    //     import(/* webpackChunkName: "markdown" */ '../fragments/Markdown.react'),
+    markdown: () =>
+        import(/* webpackChunkName: "markdown" */ '../fragments/Markdown.react'),
     plotly: () =>
         Promise.resolve(
             window.Plotly ||
@@ -23,6 +23,6 @@ export default {
     //     import(/* webpackChunkName: "slider" */ '../fragments/RangeSlider.react'),
     // slider: () =>
     //     import(/* webpackChunkName: "slider" */ '../fragments/Slider.react'),
-    // upload: () =>
-    //     import(/* webpackChunkName: "upload" */ '../fragments/Upload.react'),
+    upload: () =>
+        import(/* webpackChunkName: "upload" */ '../fragments/Upload.react'),
 };
